Move internal reader abstractions into separate file
